+++ src/db/results.rs
@@ -87,6 +87,13 @@ pub struct PoolState {
     pub idle_connections: u32,
 }
 
+#[derive(Debug, Default)]
+pub struct ConnectionInfo {
+    pub age: i64,
+    pub spanner_age: i64,
+    pub spanner_idle: i64,
+}
+
 pub type GetCollectionId = i32;
 
 #[cfg(test)]