+++ src/settings.rs
@@ -33,6 +33,13 @@ pub struct Quota {
     pub enforced: bool,
 }
 
+#[derive(Copy, Clone, Default, Debug)]
+pub struct Deadman {
+    pub max_size: Option<u32>,
+    pub previous_count: usize,
+    pub clock_start: Option<time::Instant>,
+}
+
 #[derive(Clone, Debug, Deserialize)]
 pub struct Settings {
     pub debug: bool,
@@ -43,6 +50,12 @@ pub struct Settings {
     pub database_pool_max_size: Option<u32>,
     // NOTE: Not supported by deadpool!
     pub database_pool_min_idle: Option<u32>,
+    /// Pool timeout when waiting for a slot to become available, in seconds
+    pub database_pool_connection_timeout: Option<u32>,
+    /// Max age a given connection should live, in seconds
+    pub database_pool_connection_lifespan: Option<u32>,
+    /// Max time a connection should sit idle before being dropped.
+    pub database_pool_connection_max_idle: Option<u32>,
     #[cfg(test)]
     pub database_use_test_transactions: bool,
 
@@ -79,6 +92,9 @@ impl Default for Settings {
             tokenserver_database_url: None,
             database_pool_max_size: None,
             database_pool_min_idle: None,
+            database_pool_connection_lifespan: None,
+            database_pool_connection_max_idle: None,
+            database_pool_connection_timeout: Some(30),
             #[cfg(test)]
             database_use_test_transactions: false,
             actix_keep_alive: None,
@@ -108,8 +124,16 @@ impl Settings {
         s.set_default("host", "127.0.0.1")?;
         s.set_default("human_logs", false)?;
         #[cfg(test)]
+        s.set_default("database_pool_connection_timeout", Some(30))?;
+        // Max lifespan a connection should have.
+        s.set_default::<Option<String>>("database_connection_lifespan", None)?;
+        // Max time a connection should be idle before dropping.
+        s.set_default::<Option<String>>("database_connection_max_idle", None)?;
         s.set_default("database_use_test_transactions", false)?;
         s.set_default("master_secret", "")?;
+        // Each backend does their own default process, so specifying a "universal" value
+        // for database_pool_max_size doesn't quite work. Generally the max pool size is
+        // 10.
         s.set_default::<Option<String>>("tokenserver_database_url", None)?;
         s.set_default::<Option<String>>("tokenserver_jwks_rsa_modulus", None)?;
         s.set_default::<Option<String>>("tokenserver_jwks_rsa_exponent", None)?;
@@ -342,5 +366,7 @@ pub fn test_settings() -> Settings {
     settings.port = 8000;
     settings.database_pool_max_size = Some(1);
     settings.database_use_test_transactions = true;
+    settings.database_pool_connection_max_idle = Some(300);
+    settings.database_pool_connection_lifespan = Some(300);
     settings
 }