+++ src/web/handlers.rs
@@ -1,7 +1,10 @@
 //! API Handlers
 use std::collections::HashMap;
+use std::convert::Into;
 
-use actix_web::{http::StatusCode, web::Data, Error, HttpRequest, HttpResponse};
+use actix_web::{
+    dev::HttpResponseBuilder, http::StatusCode, web::Data, Error, HttpRequest, HttpResponse,
+};
 use serde::Serialize;
 use serde_json::{json, Value};
 
@@ -29,9 +32,10 @@ pub const ONE_KB: f64 = 1024.0;
 pub async fn get_collections(
     meta: MetaRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             meta.metrics.incr("request.get_collections");
             let result = db.get_collection_timestamps(meta.user_id).await?;
 
@@ -45,9 +49,10 @@ pub async fn get_collections(
 pub async fn get_collection_counts(
     meta: MetaRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             meta.metrics.incr("request.get_collection_counts");
             let result = db.get_collection_counts(meta.user_id).await?;
 
@@ -61,9 +66,10 @@ pub async fn get_collection_counts(
 pub async fn get_collection_usage(
     meta: MetaRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             meta.metrics.incr("request.get_collection_usage");
             let usage: HashMap<_, _> = db
                 .get_collection_usage(meta.user_id)
@@ -82,9 +88,10 @@ pub async fn get_collection_usage(
 pub async fn get_quota(
     meta: MetaRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             meta.metrics.incr("request.get_quota");
             let usage = db.get_storage_usage(meta.user_id).await?;
             Ok(HttpResponse::Ok().json(vec![Some(usage as f64 / ONE_KB), None]))
@@ -95,9 +102,10 @@ pub async fn get_quota(
 pub async fn delete_all(
     meta: MetaRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             meta.metrics.incr("request.delete_all");
             Ok(HttpResponse::Ok().json(db.delete_storage(meta.user_id).await?))
         })
@@ -107,9 +115,10 @@ pub async fn delete_all(
 pub async fn delete_collection(
     coll: CollectionRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             let delete_bsos = !coll.query.ids.is_empty();
             let metrics = coll.metrics.clone();
             let timestamp: ApiResult<SyncTimestamp> = if delete_bsos {
@@ -147,14 +156,16 @@ pub async fn delete_collection(
             Ok(resp.json(timestamp))
         })
         .await
+        .map_err(Into::into)
 }
 
 pub async fn get_collection(
     coll: CollectionRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             coll.metrics.clone().incr("request.get_collection");
             let params = params::GetBsos {
                 user_id: coll.user_id.clone(),
@@ -227,9 +238,10 @@ where
 pub async fn post_collection(
     coll: CollectionPostRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             coll.metrics.clone().incr("request.post_collection");
             trace!("Collection: Post");
 
@@ -318,7 +330,6 @@ pub async fn post_collection_batch(
         .await?
     };
 
-    let commit = breq.commit;
     let user_id = coll.user_id.clone();
     let collection = coll.collection.clone();
 
@@ -326,69 +337,56 @@ pub async fn post_collection_batch(
     let mut failed = coll.bsos.invalid;
     let bso_ids: Vec<_> = coll.bsos.valid.iter().map(|bso| bso.id.clone()).collect();
 
-    let result = if commit && !coll.bsos.valid.is_empty() {
-        // There's pending items to append to the batch but since we're
-        // committing, write them to bsos immediately. Otherwise under
-        // Spanner we would pay twice the mutations for those pending
-        // items (once writing them to to batch_bsos, then again
-        // writing them to bsos)
-        trace!("Batch: Committing {}", &new_batch.id);
-        db.post_bsos(params::PostBsos {
-            user_id: coll.user_id.clone(),
-            collection: coll.collection.clone(),
-            // XXX: why does BatchBsoBody exist (it's the same struct
-            // as PostCollectionBso)?
-            bsos: coll
-                .bsos
-                .valid
-                .into_iter()
-                .map(|batch_bso| params::PostCollectionBso {
-                    id: batch_bso.id,
-                    sortindex: batch_bso.sortindex,
-                    payload: batch_bso.payload,
-                    ttl: batch_bso.ttl,
-                })
-                .collect(),
-            for_batch: true,
-            failed: Default::default(),
-        })
-        .await
-        .map(|_| ())
-    } else {
-        // We're not yet to commit the accumulated batch, but there are some
-        // additional records we need to add.
-        trace!("Batch: Appending to {}", &new_batch.id);
-        db.append_to_batch(params::AppendToBatch {
-            user_id: coll.user_id.clone(),
-            collection: coll.collection.clone(),
-            batch: new_batch.clone(),
-            bsos: coll.bsos.valid.into_iter().map(From::from).collect(),
-        })
-        .await
-    };
+    let mut resp: Value = json!({});
 
-    // collect up the successful and failed bso_ids into a response.
-    match result {
-        Ok(_) => success.extend(bso_ids),
-        Err(e) if e.is_conflict() => return Err(e.into()),
-        Err(apperr) => {
-            if let ApiErrorKind::Db(dberr) = apperr.kind() {
-                // If we're over quota, return immediately with a 403 to let the client know.
-                // Otherwise the client will simply keep retrying records.
-                if let DbErrorKind::Quota = dberr.kind() {
-                    return Err(apperr.into());
+    macro_rules! handle_result {
+            // collect up the successful and failed bso_ids into a response.
+            ( $r: expr) => {
+            match $r {
+                Ok(_) => success.extend(bso_ids.clone()),
+                Err(e) if e.is_conflict() => return Err(e.into()),
+                Err(apperr) => {
+                    if let ApiErrorKind::Db(dberr) = apperr.kind() {
+                        // If we're over quota, return immediately with a 403 to let the client know.
+                        // Otherwise the client will simply keep retrying records.
+                        if let DbErrorKind::Quota = dberr.kind() {
+                            return Err(apperr.into());
+                        }
+                    };
+                    failed.extend(
+                        bso_ids
+                            .clone()
+                            .into_iter()
+                            .map(|id| (id, "db error".to_owned())),
+                    )
                 }
             };
-            failed.extend(bso_ids.into_iter().map(|id| (id, "db error".to_owned())))
         }
-    };
+    }
 
-    let mut resp = json!({
-        "success": success,
-        "failed": failed,
-    });
-
+    // If we're not committing the current set of records yet.
     if !breq.commit {
+        // and there are bsos included in this message.
+        if !coll.bsos.valid.is_empty() {
+            // Append the data to the requested batch.
+            let result = {
+                dbg!("Batch: Appending to {}", &new_batch.id);
+                db.append_to_batch(params::AppendToBatch {
+                    user_id: coll.user_id.clone(),
+                    collection: coll.collection.clone(),
+                    batch: new_batch.clone(),
+                    bsos: coll.bsos.valid.into_iter().map(From::from).collect(),
+                })
+                .await
+            };
+            handle_result!(result);
+        }
+
+        // Return the batch append response without committing the current
+        // batch to the BSO table.
+        resp["success"] = json!(success);
+        resp["failed"] = json!(failed);
+
         resp["batch"] = json!(&new_batch.id);
         return Ok(HttpResponse::Accepted().json(resp));
     }
@@ -404,6 +402,8 @@ pub async fn post_collection_batch(
 
     // TODO: validate *actual* sizes of the batch items
     // (max_total_records, max_total_bytes)
+    //
+    // First, write the pending batch BSO data into the BSO table.
     let modified = if let Some(batch) = batch {
         db.commit_batch(params::CommitBatch {
             user_id: user_id.clone(),
@@ -416,6 +416,41 @@ pub async fn post_collection_batch(
         return Err(ApiError::from(err).into());
     };
 
+    // Then, write the BSOs contained in the commit request into the BSO table.
+    // (This presumes that the BSOs contained in the final "commit" message are
+    // newer, and thus more "correct", than any prior BSO info that may have been
+    // included in the prior batch creation messages. The client shouldn't really
+    // be including BSOs with the commit message, however it does and we should
+    // handle that case.)
+    if !coll.bsos.valid.is_empty() {
+        trace!("Batch: writing commit message bsos");
+        let result = db
+            .post_bsos(params::PostBsos {
+                user_id: coll.user_id.clone(),
+                collection: coll.collection.clone(),
+                bsos: coll
+                    .bsos
+                    .valid
+                    .into_iter()
+                    .map(|batch_bso| params::PostCollectionBso {
+                        id: batch_bso.id,
+                        sortindex: batch_bso.sortindex,
+                        payload: batch_bso.payload,
+                        ttl: batch_bso.ttl,
+                    })
+                    .collect(),
+                for_batch: false,
+                failed: Default::default(),
+            })
+            .await
+            .map(|_| ());
+
+        handle_result!(result);
+    }
+
+    // Always return success, failed, & modified
+    resp["success"] = json!(success);
+    resp["failed"] = json!(failed);
     resp["modified"] = json!(modified);
     trace!("Batch: Returning result: {}", &resp);
     Ok(HttpResponse::build(StatusCode::OK)
@@ -426,9 +461,10 @@ pub async fn post_collection_batch(
 pub async fn delete_bso(
     bso_req: BsoRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             bso_req.metrics.incr("request.delete_bso");
             let result = db
                 .delete_bso(params::DeleteBso {
@@ -445,9 +481,10 @@ pub async fn delete_bso(
 pub async fn get_bso(
     bso_req: BsoRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             bso_req.metrics.incr("request.get_bso");
             let result = db
                 .get_bso(params::GetBso {
@@ -468,9 +505,10 @@ pub async fn get_bso(
 pub async fn put_bso(
     bso_req: BsoPutRequest,
     db_pool: DbTransactionPool,
+    request: HttpRequest,
 ) -> Result<HttpResponse, Error> {
     db_pool
-        .transaction_http(|db| async move {
+        .transaction_http(request, |db| async move {
             bso_req.metrics.incr("request.put_bso");
             let result = db
                 .put_bso(params::PutBso {
@@ -538,6 +576,81 @@ pub async fn heartbeat(hb: HeartbeatRequest) -> Result<HttpResponse, Error> {
             Ok(HttpResponse::ServiceUnavailable().json(checklist))
         }
     }
+}
+
+pub async fn lbheartbeat(req: HttpRequest) -> Result<HttpResponse, Error> {
+    let mut resp: HashMap<String, Value> = HashMap::new();
+
+    let state = match req.app_data::<Data<ServerState>>() {
+        Some(s) => s,
+        None => {
+            error!("⚠️ Could not load the app state");
+            return Ok(HttpResponse::InternalServerError().body(""));
+        }
+    };
+
+    let deadarc = state.deadman.clone();
+    let mut deadman = *deadarc.read().await;
+    let db_state = if cfg!(test) {
+        use crate::db::results::PoolState;
+        use actix_web::http::header::HeaderValue;
+        use std::str::FromStr;
+
+        let test_pool = PoolState {
+            connections: u32::from_str(
+                req.headers()
+                    .get("TEST_CONNECTIONS")
+                    .unwrap_or(&HeaderValue::from_static("0"))
+                    .to_str()
+                    .unwrap_or("0"),
+            )
+            .unwrap_or_default(),
+            idle_connections: u32::from_str(
+                req.headers()
+                    .get("TEST_IDLES")
+                    .unwrap_or(&HeaderValue::from_static("0"))
+                    .to_str()
+                    .unwrap_or("0"),
+            )
+            .unwrap_or_default(),
+        };
+        // dbg!(&test_pool, deadman.max_size);
+        test_pool
+    } else {
+        state.db_pool.clone().state()
+    };
+
+    let active = db_state.connections - db_state.idle_connections;
+    let mut status_code = StatusCode::OK;
+
+    if let Some(max_size) = deadman.max_size {
+        if active >= max_size && db_state.idle_connections == 0 {
+            if deadman.clock_start.is_none() {
+                deadman.clock_start = Some(time::Instant::now());
+            }
+            status_code = StatusCode::INTERNAL_SERVER_ERROR;
+        } else if deadman.clock_start.is_some() {
+            deadman.clock_start = None
+        }
+        deadman.previous_count = db_state.idle_connections as usize;
+        {
+            *deadarc.write().await = deadman;
+        }
+        resp.insert("active_connections".to_string(), Value::from(active));
+        resp.insert(
+            "idle_connections".to_string(),
+            Value::from(db_state.idle_connections),
+        );
+        if let Some(clock) = deadman.clock_start {
+            let duration: time::Duration = time::Instant::now() - clock;
+            resp.insert(
+                "duration_ms".to_string(),
+                Value::from(duration.whole_milliseconds()),
+            );
+        };
+    }
+
+    Ok(HttpResponseBuilder::new(status_code).json(json!(resp)))
 }
 
 // try returning an API error