+++ src/web/middleware/mod.rs
@@ -1,4 +1,3 @@
-// pub mod db;
 pub mod rejectua;
 pub mod sentry;
 pub mod weave;