+++ src/web/tags.rs
@@ -1,5 +1,7 @@
+use core::cell::RefMut;
 use std::collections::{BTreeMap, HashMap};
 
+use actix_http::Extensions;
 use actix_web::{
     dev::{Payload, RequestHead},
     http::header::USER_AGENT,
@@ -52,17 +54,76 @@ fn insert_if_not_empty(label: &str, val: &str, tags: &mut HashMap<String, String
     }
 }
 
-// Tags are extra data to be recorded in metric and logging calls.
-// If additional tags are required or desired, you will need to add them to the
-// mutable extensions, e.g.
-// ```
-//      let mut tags = request.extensions_mut().get::<Tags>();
-//      tags.insert("SomeLabel".to_owned(), "whatever".to_owned());
-// ```
-// how you get the request (or the response, and it's set of `extensions`) to whatever
-// function requires it, is left as an exercise for the reader.
+/// Tags are extra data to be recorded in metric and logging calls.
+///
+/// If additional tags are required or desired, you will need to add them to the
+/// mutable extensions, e.g.
+/// ```compile_fail
+///      let mut tags = Tags::default();
+///      tags.add_tag("SomeLabel", "whatever");
+///      tags.commit(&mut request.extensions_mut());
+/// ```
 impl Tags {
-    pub fn from_request_head(req_head: &RequestHead) -> Tags {
+    pub fn extend(&mut self, new_tags: Self) {
+        self.tags.extend(new_tags.tags);
+        self.extra.extend(new_tags.extra);
+    }
+
+    pub fn with_tags(tags: HashMap<String, String>) -> Tags {
+        if tags.is_empty() {
+            return Tags::default();
+        }
+        Tags {
+            tags,
+            extra: HashMap::new(),
+        }
+    }
+
+    pub fn add_extra(&mut self, key: &str, value: &str) {
+        if !value.is_empty() {
+            self.extra.insert(key.to_owned(), value.to_owned());
+        }
+    }
+
+    pub fn add_tag(&mut self, key: &str, value: &str) {
+        if !value.is_empty() {
+            self.tags.insert(key.to_owned(), value.to_owned());
+        }
+    }
+
+    pub fn get(&self, label: &str) -> String {
+        let none = "None".to_owned();
+        self.tags.get(label).map(String::from).unwrap_or(none)
+    }
+
+    pub fn tag_tree(self) -> BTreeMap<String, String> {
+        let mut result = BTreeMap::new();
+
+        for (k, v) in self.tags {
+            result.insert(k.clone(), v.clone());
+        }
+        result
+    }
+
+    pub fn extra_tree(self) -> BTreeMap<String, Value> {
+        let mut result = BTreeMap::new();
+
+        for (k, v) in self.extra {
+            result.insert(k.clone(), Value::from(v));
+        }
+        result
+    }
+
+    pub fn commit(self, exts: &mut RefMut<'_, Extensions>) {
+        match exts.get_mut::<Tags>() {
+            Some(t) => t.extend(self),
+            None => exts.insert(self),
+        }
+    }
+}
+
+impl From<&RequestHead> for Tags {
+    fn from(req_head: &RequestHead) -> Self {
         // Return an Option<> type because the later consumers (ApiErrors) presume that
         // tags are optional and wrapped by an Option<> type.
         let mut tags = HashMap::new();
@@ -84,43 +145,6 @@ impl Tags {
         extra.insert("uri.path".to_owned(), req_head.uri.to_string());
         Tags { tags, extra }
     }
-
-    pub fn with_tags(tags: HashMap<String, String>) -> Tags {
-        if tags.is_empty() {
-            return Tags::default();
-        }
-        Tags {
-            tags,
-            extra: HashMap::new(),
-        }
-    }
-
-    pub fn get(&self, label: &str) -> String {
-        let none = "None".to_owned();
-        self.tags.get(label).map(String::from).unwrap_or(none)
-    }
-
-    pub fn extend(&mut self, tags: HashMap<String, String>) {
-        self.tags.extend(tags);
-    }
-
-    pub fn tag_tree(self) -> BTreeMap<String, String> {
-        let mut result = BTreeMap::new();
-
-        for (k, v) in self.tags {
-            result.insert(k.clone(), v.clone());
-        }
-        result
-    }
-
-    pub fn extra_tree(self) -> BTreeMap<String, Value> {
-        let mut result = BTreeMap::new();
-
-        for (k, v) in self.extra {
-            result.insert(k.clone(), Value::from(v));
-        }
-        result
-    }
 }
 
 impl FromRequest for Tags {
@@ -133,7 +157,7 @@ impl FromRequest for Tags {
             let exts = req.extensions();
             match exts.get::<Tags>() {
                 Some(t) => t.clone(),
-                None => Tags::from_request_head(req.head()),
+                None => Tags::from(req.head()),
             }
         };