+++ CHANGELOG.md
@@ -1,3 +1,99 @@
+<a name="0.9.1"></a>
+## 0.9.1 (2021-03-12)
+
+
+#### Chore
+
+*   kill the long unused db middleware ([8f9cce76](https://github.com/mozilla-services/syncstorage-rs/commit/8f9cce76ba4a52e4594f32b471f2e0259abe04d2), closes [#693](https://github.com/mozilla-services/syncstorage-rs/issues/693), [#1018](https://github.com/mozilla-services/syncstorage-rs/issues/1018))
+*   Update for March 2021 ([4e38e681](https://github.com/mozilla-services/syncstorage-rs/commit/4e38e68180766c083b651d148c24f42e5d0fd058), closes [#1018](https://github.com/mozilla-services/syncstorage-rs/issues/1018))
+
+#### Features
+
+*   Manually update the spanner session approximate_last_used_time (#1009) ([f669b257](https://github.com/mozilla-services/syncstorage-rs/commit/f669b257a2d8b8f4928a32df27eebe33f1af555e), closes [#1008](https://github.com/mozilla-services/syncstorage-rs/issues/1008))
+
+
+
+<a name="0.9.0"></a>
+## 0.9.0 (2021-02-25)
+
+
+#### Bug Fixes
+
+*   ensure "extra" data added to Hawk header in e2e tests (#1003) ([8c280ccd](https://github.com/mozilla-services/syncstorage-rs/commit/8c280ccda032ff007c4a6017d6fb0cdd94d7cd3f))
+
+#### Features
+
+*   kill old or excessively idled connections (#1006) ([082dd1f0](https://github.com/mozilla-services/syncstorage-rs/commit/082dd1f0613fc91f3ea2f02b3bcdd9ddf2b938d3))
+
+#### Chore
+
+*   fmt & clippy fixes for Rust 1.50 (#1004) ([56cadcb2](https://github.com/mozilla-services/syncstorage-rs/commit/56cadcb2cdcce99de2d425e8d0edb4a7c20c52ac))
+*   RUSTSEC-2021-0020 fix ([2e186341](https://github.com/mozilla-services/syncstorage-rs/commit/2e1863410ed92180f6fb12a9ebf8d2f462425b38), closes [#999](https://github.com/mozilla-services/syncstorage-rs/issues/999))
+*   tag 0.8.7 (#998) ([d06b3c2a](https://github.com/mozilla-services/syncstorage-rs/commit/d06b3c2a0dd1602d074d5d2da913db699eea0a9e))
+
+
+
+<a name="0.8.7"></a>
+### 0.8.7  (2021-02-03)
+
+
+#### Features
+
+*   `__lbheartbeat__` will return 500 if the connection pool is exhausted (#997) ([e72573ac](https://github.com/mozilla-services/syncstorage-rs/commit/e72573acedce2916c9fd3aa8e3c54fbe71f2008e), closes [#996](https://github.com/mozilla-services/syncstorage-rs/issues/996))
+
+#### Chore
+
+*   tag 0.8.6 (#995) ([8cb5b603](https://github.com/mozilla-services/syncstorage-rs/commit/8cb5b603f0320483904107eee682797b8d814a44))
+
+
+
+<a name="0.8.6"></a>
+### 0.8.6 (2021-02-01)
+
+
+#### Refactor
+
+*   remove duplicate code for incrementing counters (#983) ([d72228b1](https://github.com/mozilla-services/syncstorage-rs/commit/d72228b1d4b5cd63a399bde77c3156ea53bb4217))
+
+#### Chore
+
+*   tag 0.8.5 (#979) ([3c23fb46](https://github.com/mozilla-services/syncstorage-rs/commit/3c23fb46138d4a042d0293af6b9853ea9f173f6d))
+
+#### Features
+
+*   Add pool connection info to __lbheartbeat__ for ops (#985) ([06a2ac79](https://github.com/mozilla-services/syncstorage-rs/commit/06a2ac7910a87a75f1a2f0d68e786579cec99fd8))
+
+#### Bug Fixes
+
+*   downgrade deadpool so it stays on tokio 0.2 ([99975ef8](https://github.com/mozilla-services/syncstorage-rs/commit/99975ef8b64317511111d48c6ebfc75e7facc334), closes [#976](https://github.com/mozilla-services/syncstorage-rs/issues/976))
+
+
+
+<a name="0.8.5"></a>
+## 0.8.5 (2021-01-21)
+
+
+#### Bug Fixes
+
+*   downgrade deadpool so it stays on tokio 0.2 ([99975ef8](https://github.com/mozilla-services/syncstorage-rs/commit/99975ef8b64317511111d48c6ebfc75e7facc334), closes [#976](https://github.com/mozilla-services/syncstorage-rs/issues/976))
+
+
+
+<a name="0.8.4"></a>
+### 0.8.4 (2021-01-19)
+
+
+#### Chore
+*   Update pyo3 to the latest version (#938) ([cc7d9d36]https://github.com/mozilla-services/syncstorage-rs/commit/cc7d9d367310aeb7551668c049f1a895a6eae853))
+*   update dependencies (#953) ([bca8770f](https://github.com/mozilla-services/syncstorage-rs/commit/bca8770f531b45b00e57e137082b1ed9d90acd7f))
+*   tag 0.8.3 (#937) ([02b76231](https://github.com/mozilla-services/syncstorage-rs/commit/02b76231cf4fa015093cea75286a82f306c833b0))
+
+
+#### Features
+
+*   default to timing out deadpool checkouts (30 seconds) (#974) ([2ecca202](https://github.com/mozilla-services/syncstorage-rs/commit/2ecca202aa01f123898115827af6e5967f8a1e9b), closes [#973](https://github.com/mozilla-services/syncstorage-rs/issues/973))
+*   avoid an unnecessarily cloning for from_raw_bso (#972) ([07352b6d](https://github.com/mozilla-services/syncstorage-rs/commit/07352b6d7a331d07e18ec386a650d3b720c5703f), closes [#971](https://github.com/mozilla-services/syncstorage-rs/issues/971))
+
 <a name="0.8.3"></a>
 ### 0.8.3 (2020-11-30)
 
@@ -5,7 +101,6 @@
 #### Chore
 
 *   Update to rust 1.48 (#927) ([ea1f222b](https://github.com/mozilla-services/syncstorage-rs/commit/ea1f222b219ddd78684945058c3b3430ed636982))
-*   Update pyo3 to the latest version (#938) ([cc7d9d36]https://github.com/mozilla-services/syncstorage-rs/commit/cc7d9d367310aeb7551668c049f1a895a6eae853))
 
 <a name="0.8.2"></a>
 ## 0.8.2 (2020-11-20)